+++ novawallet/Common/Protocols/BaseErrorPresentable.swift
@@ -21,6 +21,12 @@ protocol BaseErrorPresentable {
     )
 
     func presentMinBalanceViolated(from view: ControllerBackedProtocol, locale: Locale?)
+
+    func presentNotValidAddress(
+        from view: ControllerBackedProtocol,
+        networkName: String,
+        locale: Locale?
+    )
 }
 
 extension BaseErrorPresentable where Self: AlertPresentable & ErrorPresentable {
@@ -167,6 +173,23 @@ extension BaseErrorPresentable where Self: AlertPresentable & ErrorPresentable {
         )
 
         let closeAction = R.string.localizable.commonClose(preferredLanguages: locale?.rLanguages)
+
+        present(message: message, title: title, closeAction: closeAction, from: view)
+    }
+
+    func presentNotValidAddress(
+        from view: ControllerBackedProtocol,
+        networkName: String,
+        locale: Locale?
+    ) {
+        let title = R.string.localizable.commonErrorInvalidAddressTitle(
+            preferredLanguages: locale?.rLanguages)
+        let message = R.string.localizable.commonErrorInvalidAddressMessage(
+            networkName,
+            preferredLanguages: locale?.rLanguages
+        )
+        let closeAction = R.string.localizable.commonClose(
+            preferredLanguages: locale?.rLanguages)
 
         present(message: message, title: title, closeAction: closeAction, from: view)
     }
