+++ novawallet/Common/Validation/Validators/BaseDataValidatorFactory.swift
@@ -37,6 +37,12 @@ protocol BaseDataValidatingFactoryProtocol: AnyObject {
         minBalance: BigUInt?,
         locale: Locale
     ) -> DataValidating
+
+    func validAddress(
+        _ address: String,
+        chain: ChainModel,
+        locale: Locale
+    ) -> DataValidating
 }
 
 extension BaseDataValidatingFactoryProtocol {
@@ -261,6 +267,26 @@ extension BaseDataValidatingFactoryProtocol {
             } else {
                 return false
             }
+        })
+    }
+
+    func validAddress(
+        _ address: String,
+        chain: ChainModel,
+        locale: Locale
+    ) -> DataValidating {
+        ErrorConditionViolation(onError: { [weak self] in
+            guard let view = self?.view else {
+                return
+            }
+            self?.basePresentable.presentNotValidAddress(
+                from: view,
+                networkName: chain.name,
+                locale: locale
+            )
+        }, preservesCondition: {
+            let accountId = try? address.toAccountId(using: chain.chainFormat)
+            return accountId != nil
         })
     }
 }
