+++ novawallet/Modules/Staking/StakingSetupProxy/Validation/ProxyDataValidatorFactory.swift
@@ -4,16 +4,10 @@ import SoraFoundation
 
 protocol ProxyDataValidatorFactoryProtocol: BaseDataValidatingFactoryProtocol {
     func hasSufficientBalance(
-        available: BigUInt?,
+        available: BigUInt,
         deposit: BigUInt?,
         fee: BigUInt?,
         asset: AssetBalanceDisplayInfo,
-        locale: Locale
-    ) -> DataValidating
-
-    func validAddress(
-        _ address: String,
-        chain: ChainModel,
         locale: Locale
     ) -> DataValidating
 
@@ -49,7 +43,7 @@ final class ProxyDataValidatorFactory: ProxyDataValidatorFactoryProtocol {
     }
 
     func hasSufficientBalance(
-        available: BigUInt?,
+        available: BigUInt,
         deposit: BigUInt?,
         fee: BigUInt?,
         asset: AssetBalanceDisplayInfo,
@@ -61,7 +55,7 @@ final class ProxyDataValidatorFactory: ProxyDataValidatorFactoryProtocol {
                 let viewModelFactory = self?.balanceViewModelFactoryFacade else {
                 return
             }
-            let balanceDecimal = available?.decimal(assetInfo: asset)
+            let balanceDecimal = available.decimal(assetInfo: asset)
             let depositDecimal = deposit?.decimal(assetInfo: asset)
 
             let balanceModel = viewModelFactory.amountFromValue(
@@ -81,31 +75,10 @@ final class ProxyDataValidatorFactory: ProxyDataValidatorFactoryProtocol {
             )
         }, preservesCondition: {
             guard let deposit = deposit,
-                  let fee = fee,
-                  let available = available else {
+                  let fee = fee else {
                 return false
             }
             return available >= deposit + fee
-        })
-    }
-
-    func validAddress(
-        _ address: String,
-        chain: ChainModel,
-        locale: Locale
-    ) -> DataValidating {
-        ErrorConditionViolation(onError: { [weak self] in
-            guard let view = self?.view else {
-                return
-            }
-            self?.presentable.presentNotValidAddress(
-                from: view,
-                networkName: chain.name,
-                locale: locale
-            )
-        }, preservesCondition: {
-            let accountId = try? address.toAccountId(using: chain.chainFormat)
-            return accountId != nil
         })
     }
 
@@ -133,7 +106,7 @@ final class ProxyDataValidatorFactory: ProxyDataValidatorFactoryProtocol {
                   let limit = limit else {
                 return false
             }
-            return proxyCount < limit
+            return proxyCount <= limit
         })
     }
