+++ novawallet/Modules/Transfer/TransferSetup/TransferSetupViewFactory.swift
@@ -163,8 +163,9 @@ enum TransferSetupViewFactory {
             for: nil,
             sortDescriptors: [NSSortDescriptor.accountsByOrder]
         )
+        let operationQueue = OperationManagerFacade.sharedDefaultQueue
 
-        let web3NameService = createWeb3NameService()
+        let web3NameService = Web3NameServiceFactory(operationQueue: operationQueue).createService()
 
         return TransferSetupInteractor(
             chainAsset: params.chainAsset,
@@ -175,38 +176,6 @@ enum TransferSetupViewFactory {
             accountRepository: accountRepository,
             web3NamesService: web3NameService,
             operationManager: OperationManager()
-        )
-    }
-
-    private static func createWeb3NameService() -> Web3NameServiceProtocol? {
-        let kiltChainId = KnowChainId.kiltOnEnviroment
-        let chainRegistry = ChainRegistryFacade.sharedRegistry
-
-        guard let kiltConnection = chainRegistry.getConnection(for: kiltChainId),
-              let kiltRuntimeService = chainRegistry.getRuntimeProvider(for: kiltChainId) else {
-            return nil
-        }
-
-        let operationQueue = OperationManagerFacade.sharedDefaultQueue
-        let web3NamesOperationFactory = KiltWeb3NamesOperationFactory(operationQueue: operationQueue)
-
-        let recipientRepositoryFactory = Web3TransferRecipientRepositoryFactory(
-            integrityVerifierFactory: Web3TransferRecipientIntegrityVerifierFactory()
-        )
-
-        let slip44CoinsUrl = ApplicationConfig.shared.slip44URL
-        let slip44CoinsProvider: AnySingleValueProvider<Slip44CoinList> = JsonDataProviderFactory.shared.getJson(
-            for: slip44CoinsUrl
-        )
-
-        return Web3NameService(
-            providerName: Web3NameProvider.kilt,
-            slip44CoinsProvider: slip44CoinsProvider,
-            web3NamesOperationFactory: web3NamesOperationFactory,
-            runtimeService: kiltRuntimeService,
-            connection: kiltConnection,
-            transferRecipientRepositoryFactory: recipientRepositoryFactory,
-            operationQueue: operationQueue
         )
     }
 }
