+++ novawallet/en.lproj/Localizable.strings
@@ -1380,8 +1380,8 @@
 "staking.setup.add.your.proxy" = "Add delegated authority (Proxy)";
 "staking.setup.proxy.error.insufficient.balance.title" = "Not enough tokens";
 "staking.setup.proxy.error.insufficient.balance.message" = "You don’t have enough balance for proxy deposit of %@. Available balance: %@";
-"staking.setup.proxy.error.invalid.address.title" = "Invalid proxy address";
-"staking.setup.proxy.error.invalid.address.message" = "Proxy address should be a valid %@ address";
+"common.error.invalid.address.title" = "Invalid address";
+"common.error.invalid.address.message" = "Address should be a valid %@ address";
 "staking.setup.proxy.error.invalid.maximum.proxies.title" = "Maximum number of proxies has been reached";
 "staking.setup.proxy.error.invalid.maximum.proxies.message" = "You have reached the limit of %@ added proxies in %@. Remove proxies to add new ones.";
 "staking.setup.proxy.error.proxy.already.exists.title" = "Delegation already exists";
