+++ novawallet/ru.lproj/Localizable.strings
@@ -1380,8 +1380,8 @@
 "staking.setup.add.your.proxy" = "Добавить делегацию (Прокси)";
 "staking.setup.proxy.error.insufficient.balance.title" = "Недостаточно токенов";
 "staking.setup.proxy.error.insufficient.balance.message" = "У вас недостаточно средств для депозита прокси %@. Доступный баланс: %@";
-"staking.setup.proxy.error.invalid.address.title" = "Некорректный адрес прокси";
-"staking.setup.proxy.error.invalid.address.message" = "Адрес прокси должен быть допустимым адресом в сети %@";
+"common.error.invalid.address.title" = "Некорректный адрес";
+"common.error.invalid.address.message" = "Адрес должен быть допустимым адресом в сети %@";
 "staking.setup.proxy.error.invalid.maximum.proxies.title" = "Достигнуто максимальное количество прокси";
 "staking.setup.proxy.error.invalid.maximum.proxies.message" = "Вы достигли лимита добавленных прокси (%@) в %@. Чтобы добавить новые, удалите существующие.";
 "staking.setup.proxy.error.proxy.already.exists.title" = "Прокси уже добавлен";
